requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
pyaudio==0.2.14
websockets==12.0
python-linkedin-v2==0.8.6
chromadb==0.4.22
autogen==0.2.12
retrieve-chat==0.1.0
numpy==1.26.4
sounddevice==0.4.6
pydub==0.25.1
python-dateutil==2.8.2
pytz==2024.1 